Exposing GetTestCaseFilter from discovery context
